Split tests into more files
Move specific commands into own files.

resolves #9
